# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizPortal Multi-Division API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL for uploading a file to object storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL and final object path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
...
components:
  schemas:
...
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required: [name, size, contentType]
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
          required: [name, size, contentType]
      required: [uploadURL, objectPath, metadata]
...